pyproject.toml
[project]
name = "nemos-dream"
version = "0.1.0"
description = "EN→KR cultural SDG pipeline for NVIDIA Nemotron Hackathon Track C."
requires-python = ">=3.11"
authors = [
    { name = "nemos_dream team" },
]
readme = "README.md"

dependencies = [
    # base
    "pydantic>=2.6",
    "python-dotenv",
    "pyyaml",
    "gdown>=5.2",
    "tqdm",
    "numpy",
    "pandas>=2.0",
    # NVIDIA Nemotron stack (every stage can use any of these)
    "openai>=1.40",                        # NIM is OpenAI-compatible
    "data-designer",                       # NeMo Data Designer (stage 1)
    "nvidia-nat[langchain]",               # NeMo Agent Toolkit (stage 1)
    "nemo-curator>=1.1.0",                 # Curator pipeline + dedup (stage 3)
    "langchain-nvidia-ai-endpoints>=0.3",  # LangChain ↔ NVIDIA NIM
    "datasets",                            # HF — `nvidia/Nemotron-Personas-Korea` (stage 2)
    "tavily-python",                       # web fallback for stage 1 cultural_map (dict→retriever→web+llm)
    # Curator runtime deps (imported by `nemo_curator.stages.text.filters`)
    "ftfy",                                # UnicodeReformatter dep
    "beautifulsoup4",                      # code.py filter dep
    "justext",                             # boilerplate filter dep
    "fasttext-wheel",                      # FastTextLangId/FastTextQualityFilter dep
    "sentencepiece",                       # tokenizer for Curator code filter
    # PII gate (Curator 1.1 dropped its PiiModifier; we call Presidio directly)
    "presidio-analyzer",
    "presidio-anonymizer",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nemos_dream"]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
